CURB-4492 Further refactoring of the Filter page content
- refactored filter page in theme-gmd to use content component from @shopgate/engage/filter
- removed filter page content components from theme-gmd
- refactored FilterModal to use the FilterPage context
- added portals around apply and reset button
