+++ libraries/commerce/filter/constants/Portals.js
@@ -4,15 +4,37 @@ const FILTER = 'filter';
 // CONTENTS
 const PRICE_RANGE = 'price-range';
 const SELECTOR = 'selector';
+const APPLY_BUTTON = 'apply-button';
+const RESET_BUTTON = 'reset-button';
+const PAGE_CONTENT = 'page-content';
 
 // POSITIONS
 const BEFORE = 'before';
 const AFTER = 'after';
 
+/** @deprecated Use PORTAL_FILTER_PRICE_RANGE instead  */
 export const FILTER_PRICE_RANGE = `${FILTER}.${PRICE_RANGE}`;
+/** @deprecated Use PORTAL_FILTER_PRICE_RANGE_BEFORE instead  */
 export const FILTER_PRICE_RANGE_BEFORE = `${FILTER}.${PRICE_RANGE}.${BEFORE}`;
+/** @deprecated Use PORTAL_FILTER_PRICE_RANGE_AFTER instead  */
 export const FILTER_PRICE_RANGE_AFTER = `${FILTER}.${PRICE_RANGE}.${AFTER}`;
 
-export const FILTER_SELECTOR = `${FILTER}.${SELECTOR}`;
-export const FILTER_SELECTOR_BEFORE = `${FILTER}.${SELECTOR}.${BEFORE}`;
-export const FILTER_SELECTOR_AFTER = `${FILTER}.${SELECTOR}.${AFTER}`;
+export const PORTAL_FILTER_PRICE_RANGE = FILTER_PRICE_RANGE;
+export const PORTAL_FILTER_PRICE_RANGE_BEFORE = FILTER_PRICE_RANGE_BEFORE;
+export const PORTAL_FILTER_PRICE_RANGE_AFTER = FILTER_PRICE_RANGE_AFTER;
+
+export const PORTAL_FILTER_SELECTOR = `${FILTER}.${SELECTOR}`;
+export const PORTAL_FILTER_SELECTOR_BEFORE = `${FILTER}.${SELECTOR}.${BEFORE}`;
+export const PORTAL_FILTER_SELECTOR_AFTER = `${FILTER}.${SELECTOR}.${AFTER}`;
+
+export const PORTAL_FILTER_APPLY_BUTTON = `${FILTER}.${APPLY_BUTTON}`;
+export const PORTAL_FILTER_APPLY_BUTTON_BEFORE = `${FILTER}.${APPLY_BUTTON}.${BEFORE}`;
+export const PORTAL_FILTER_APPLY_BUTTON_AFTER = `${FILTER}.${APPLY_BUTTON}.${AFTER}`;
+
+export const PORTAL_FILTER_RESET_BUTTON = `${FILTER}.${RESET_BUTTON}`;
+export const PORTAL_FILTER_RESET_BUTTON_BEFORE = `${FILTER}.${RESET_BUTTON}.${BEFORE}`;
+export const PORTAL_FILTER_RESET_BUTTON_AFTER = `${FILTER}.${RESET_BUTTON}.${AFTER}`;
+
+export const PORTAL_FILTER_PAGE_CONTENT = `${FILTER}.${PAGE_CONTENT}`;
+export const PORTAL_FILTER_PAGE_CONTENT_BEFORE = `${FILTER}.${PAGE_CONTENT}.${BEFORE}`;
+export const PORTAL_FILTER_PAGE_CONTENT_AFTER = `${FILTER}.${PAGE_CONTENT}.${AFTER}`;
