+++ libraries/engage/filter/components/FilterPageContent/components/ApplyButton/index.jsx
@@ -1,6 +1,7 @@
 import React, { memo } from 'react';
 import PropTypes from 'prop-types';
-import { I18n, Button } from '@shopgate/engage/components';
+import { I18n, Button, SurroundPortals } from '@shopgate/engage/components';
+import { PORTAL_FILTER_APPLY_BUTTON } from '@shopgate/engage/filter/constants';
 import { withWidgetSettings } from '@shopgate/engage/core';
 import styles from './style';
 
@@ -13,18 +14,23 @@ const FilterApplyButton = ({ active, onClick, widgetSettings }) => {
   const buttonColor = active ? buttonTextColor : buttonTextColorDisabled;
 
   return (
-    <div className={styles.wrapper}>
-      <Button
-        className={styles.button}
-        flat
-        type="primary"
-        onClick={onClick}
-        disabled={!active}
-        testId="applyFilterButton"
-      >
-        <I18n.Text string="filter.apply" style={{ color: buttonColor }} />
-      </Button>
-    </div>
+    <SurroundPortals
+      portalName={PORTAL_FILTER_APPLY_BUTTON}
+      portalProps={{ active, onClick, widgetSettings }}
+    >
+      <div className={styles.wrapper}>
+        <Button
+          className={styles.button}
+          flat
+          type="primary"
+          onClick={onClick}
+          disabled={!active}
+          testId="applyFilterButton"
+        >
+          <I18n.Text string="filter.apply" style={{ color: buttonColor }} />
+        </Button>
+      </div>
+    </SurroundPortals>
   );
 };
