+++ libraries/engage/filter/components/FilterPageContent/components/ResetButton/index.jsx
@@ -1,6 +1,7 @@
 import React, { memo } from 'react';
 import PropTypes from 'prop-types';
-import { I18n, Button } from '@shopgate/engage/components';
+import { I18n, Button, SurroundPortals } from '@shopgate/engage/components';
+import { PORTAL_FILTER_RESET_BUTTON } from '@shopgate/engage/filter/constants';
 import styles from './style';
 
 /**
@@ -9,17 +10,22 @@ import styles from './style';
  * @returns {JSX}
  */
 const FilterResetButton = ({ active, onClick }) => (
-  <div className={styles}>
-    <Button
-      flat
-      type="primary"
-      onClick={onClick}
-      disabled={!active}
-      testId="clearAllButton"
-    >
-      <I18n.Text string="filter.reset" />
-    </Button>
-  </div>
+  <SurroundPortals
+    portalName={PORTAL_FILTER_RESET_BUTTON}
+    portalProps={{ active, onClick }}
+  >
+    <div className={styles}>
+      <Button
+        flat
+        type="primary"
+        onClick={onClick}
+        disabled={!active}
+        testId="clearAllButton"
+      >
+        <I18n.Text string="filter.reset" />
+      </Button>
+    </div>
+  </SurroundPortals>
 );
 
 FilterResetButton.propTypes = {
