+++ libraries/engage/filter/providers/index.js
@@ -1,1 +1,2 @@
 export { default as SortProvider } from './SortProvider';
+export { default as FilterPageProvider } from './FilterPageProvider';
