+++ themes/theme-gmd/components/FilterBar/FilterBarProvider.jsx
@@ -1,8 +1,6 @@
-import React, { useMemo, useState, useEffect } from 'react';
+import React, { useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
-import { router } from '@shopgate/engage/core';
 import { responsiveCondition } from '@shopgate/engage/styles';
-import { buildUpdatedFilters } from '@shopgate/engage/filter';
 import Context from './FilterBarProvider.context';
 import connect from './FilterBarProvider.connector';
 
@@ -18,13 +16,9 @@ const FilterBarProvider = ({
   updateFilters,
 }) => {
   const [filterModalOpen, setFilterModalOpen] = useState(false);
-  const [activeFilters, setActiveFilters] = useState(routeFilterState);
-
-  useEffect(() => {
-    setActiveFilters(routeFilterState);
-  }, [routeFilterState]);
 
   const contextValue = useMemo(() => ({
+    updateFilters,
     /**
      * Opens the filter page or modal
      * @returns {Object}
@@ -35,32 +29,10 @@ const FilterBarProvider = ({
       }
       return navigate();
     },
-    /**
-     * Applies filter to given route.
-     * @param {string} id Route id.
-     * @param {Object} filters Newly set filters.
-     * @param {Object} currentFilters Default filters.
-     * @returns {Object}
-     */
-    applyFilters: (id, filters, currentFilters) => {
-      const newFilters = buildUpdatedFilters(currentFilters, filters);
-      setFilterModalOpen(false);
-      router.update(id, { filters: newFilters });
-      return updateFilters(newFilters);
-    },
-    /**
-     * Resets the filters on given route.
-     * @param {string} id Route id.
-     * @returns {Object}
-     */
-    resetFilters: () => {
-      setActiveFilters({});
-      return updateFilters({});
-    },
-    filters: activeFilters,
+    filters: routeFilterState,
     filterModalOpen,
     setFilterModalOpen,
-  }), [activeFilters, filterModalOpen, navigate, updateFilters]);
+  }), [filterModalOpen, navigate, routeFilterState, updateFilters]);
 
   return (
     <Context.Provider value={contextValue}>
