+++ themes/theme-gmd/components/FilterBar/components/FilterModal/FilterModal.jsx
@@ -1,16 +1,12 @@
 import React, {
-  useState, useCallback, useContext,
+  useCallback, useContext,
 } from 'react';
 import { css } from 'glamor';
 import { Modal, Backdrop } from '@shopgate/engage/components';
-import { responsiveMediaQuery } from '@shopgate/engage/styles';
-import { themeConfig } from '@shopgate/pwa-common/helpers/config';
 import { RouteContext } from '@shopgate/pwa-common/context';
+import { FilterPageProvider } from '@shopgate/engage/filter/providers';
 import { useFilterBarContext } from '../../FilterBarProvider.context';
-import FilterContent from '../../../../pages/Filter/components/Content';
-import Title from './FilterModalTitle';
-
-const { colors } = themeConfig;
+import Content from './FilterModalContent';
 
 const styles = {
   root: css({
@@ -19,25 +15,6 @@ const styles = {
     alignItems: 'center',
     width: '100%',
     height: '100%',
-  }),
-  content: css({
-    zIndex: 1,
-    background: colors.light,
-    borderRadius: 4,
-    minWidth: 300,
-    [responsiveMediaQuery('>=md')]: {
-      minWidth: 600,
-    },
-    maxWidth: '50%',
-    minHeight: 300,
-    maxHeight: '90%',
-    display: 'flex',
-    flexDirection: 'column',
-    overflow: 'hidden',
-  }),
-  scrollable: css({
-    overflow: 'auto',
-    flex: 1,
   }),
 };
 
@@ -50,8 +27,7 @@ const FilterModal = () => {
     filterModalOpen,
     setFilterModalOpen,
     filters,
-    applyFilters,
-    resetFilters,
+    updateFilters,
   } = useFilterBarContext();
 
   const {
@@ -64,21 +40,14 @@ const FilterModal = () => {
     } = {},
   } = useContext(RouteContext);
 
-  const [filterState, setFilterState] = useState(null);
-
-  /* eslint-disable react-hooks/exhaustive-deps */
-  const handleApply = useCallback(() => {
-    applyFilters(id, filterState.filters, filterState.currentFilters);
-  }, [applyFilters, filterState?.currentFilters, filterState?.filters, id]);
-  /* eslint-enable react-hooks/exhaustive-deps */
-
-  const handleReset = useCallback(() => {
-    resetFilters(id);
-  }, [id, resetFilters]);
-
   const handleClose = useCallback(() => {
     setFilterModalOpen(false);
   }, [setFilterModalOpen]);
+
+  const handleApply = useCallback((update) => {
+    handleClose();
+    updateFilters(update);
+  }, [handleClose, updateFilters]);
 
   if (!filterModalOpen) {
     return null;
@@ -93,23 +62,16 @@ const FilterModal = () => {
           opacity={38}
           onClick={() => setFilterModalOpen(false)}
         />
-        <div className={styles.content}>
-          <Title
-            apply={handleApply}
-            reset={handleReset}
-            close={handleClose}
-          />
-          <div className={styles.scrollable}>
-            <FilterContent
-              parentId="0"
-              showAppBar={false}
-              onChange={setFilterState}
-              activeFilters={filters}
-              categoryId={categoryId}
-              searchPhrase={searchPhrase}
-            />
-          </div>
-        </div>
+
+        <FilterPageProvider
+          parentRouteId={id}
+          activeFilters={filters}
+          categoryId={categoryId}
+          searchPhrase={searchPhrase}
+          onApply={handleApply}
+        >
+          <Content onClose={handleClose} />
+        </FilterPageProvider>
       </div>
     </Modal>
   );
