--- themes/theme-gmd/pages/Filter/components/Content/components/ApplyButton/style.js
@@ -1,5 +1,0 @@
-import { css } from 'glamor';
-
-export default css({
-  marginTop: 8,
-});
