--- themes/theme-gmd/pages/Filter/components/Content/components/ResetButton/style.js
@@ -1,7 +1,0 @@
-import { css } from 'glamor';
-
-export default css({
-  marginTop: 4,
-  marginBottom: 40,
-  textAlign: 'right',
-});
