--- themes/theme-gmd/pages/Filter/components/Content/components/Selector/style.js
@@ -1,6 +1,0 @@
-import { css } from 'glamor';
-
-export const content = css({
-  marginLeft: -8,
-  marginBottom: -8,
-});
